% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{db_connect}
\alias{db_connect}
\title{Wrappers to easily facilitate connecting to and querying from databases.}
\usage{
db_connect(connection)
}
\arguments{
\item{connection}{String of the short-hand name of the desired database to which to connect}
}
\value{
Connection object
}
\description{
Database connections are saved as 'con' objects to your environment with db_connect() (the argument is a pre-determined short-hand so that you do not need to pass entire server strings). When querying, all you need to do is reference the 'con', and pass the SQL syntax! Make sure to assign your query output to a variable so that you can continue to manipulate it in-session.
}
\details{
Currently, this function skeleton is catered to accommodate IDOH servers. The supported short-hand database names are as follows:\tabular{ll}{
   short_hand \tab server_string \cr
   a \tab x \cr
   b \tab y \cr
   c \tab z \cr
   d \tab w \cr
   e \tab v \cr
}


For servers requiring authentication (e.g. Oracle), you will need to set up \code{keyring} tokens prior to using these functions. Simply enter \code{keyring::key_set(service, username)} where \code{service} is the short-hand name of the desired database. You will then be prompted to enter your password. See \code{?keyring::key_set} for more information.
}
\examples{
db_connect("my_fav_database")
}
