% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_tables.r
\name{fancy_table}
\alias{fancy_table}
\title{Format tabular data into a nice interactive HTML table.}
\usage{
fancy_table(df, title = quote())
}
\arguments{
\item{df}{Dataframe of the data you want fancied}

\item{title}{String of the title you want to give to your table}
}
\value{
A formatted, interactive HTML table
}
\description{
This version supports column filters, a search bar, and a scrollable Y to flexibly fit in most Markdown documents.
}
\examples{
fancy_table(mpg, title = "My Nice HTML Table")
}
