% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.r
\name{\%notin\%}
\alias{\%notin\%}
\title{Negation of \%in\% operator.}
\usage{
... \%notin\% NA
}
\value{
The complement of \%in\%
}
\description{
Yeah, that's really all it is. This seems way more intuitive
}
\examples{
4 \%notin\% c(1, 2, 3)
[1] TRUE
}
