% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.r
\name{query}
\alias{query}
\title{Convenient function to query a database connection.}
\usage{
query(con, sql_query, df_type = "tibble")
}
\arguments{
\item{con}{R connection object created using \code{icebreakr::db_connect()} or DBI/odbc}

\item{sql_query}{String of your SQL query}

\item{df_type}{String of the desired df data type output. Defaults to a tibble. "dt" converts to a data.table}
}
\value{
Tibble of the query results from the database
}
\description{
Use an in-session database connection, such as from \code{icebreakr::db_connect} or another DBI/odbc method, to pull query results directly into your R session straight from the database.
}
\examples{
x <- query(con, "select top 1000 * from my_fav_table")
}
