% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{read_clippy}
\alias{read_clippy}
\title{Convenient wrapper to get data from your clipboard into your R session.}
\usage{
read_clippy(expected_result = "df")
}
\arguments{
\item{expected_result}{String of desired data type to be read into session. Currently, only \code{df} and \code{vector} are supported. Defaults to \code{df}}
}
\value{
A dataframe or vector, depending on dimension of the data being read in
}
\description{
Works well when needing to quickly move small data from Excel into R.
}
\examples{
x <- read_clippy()
}
