% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{write_clippy}
\alias{write_clippy}
\title{Transfer R data from your environment to your clipboard.}
\usage{
write_clippy(df, col_names = T)
}
\arguments{
\item{df}{Tabular data that you want to write to clipboard}

\item{col_names}{Bool. Do you want column names written to the clipboard as well? Defaults to TRUE}
}
\value{
Clipboard buffer updates with the data passed to this function.
}
\description{
Works well when needing to quickly move small data from R into Excel.
}
\examples{
x <- mpg
write_clippy(x)
}
